feat: R9-A — per-tri IOR lane + rough-glass GI, native disc emitters on pt-webgpu (ledger → native), lint backlog ZERO + blocking CI
- Per-tri IOR (B1-ior-per-tri closed): bvh_material bits[15:8] carry quantized IOR
  [1.0,3.0] step 0.0078 (water→TiO₂); glass walk Snell/TIR + lo_transmittedGI
  Fresnel F0 now per-triangle (fixed 1.5/0.04 constants gone); flows through the
  material mutation fast path automatically; default-1.5 scenes shift < 0.0002 F0.
  BONUS: rough-glass GI — panes with roughness > 0.1 perturb the refracted walk by
  a GGX-distributed offset (frosted glass gets blurred GI); smooth path exact
- pt-webgpu disc-area NATIVE (the 32-triangle fan is gone): disc records ride the
  rect stream with a shape tag (slot 15); concentric-disc sampling (Shirley-Chiu) +
  ray-plane/radius MIS intersection + π·r² area in light-tree leaves; ALL consumers
  updated (kernel NEE, connect BSDF-half, restirPtProducer, BDPT emitter pick
  CPU+WGSL, lite tier); ledger disc-area approximate→native; new pt/disc-light
  behavioral config PASSES (gate now 29 configs)
- LINT ZERO: ~425 errors → 0 (auto-fix + per-bucket review; behavior-preserving —
  full suite as the guard; systematic test-file rule adjustments documented in
  eslint.config.js; explicit disable-with-reason where lint fights the WGSL
  template-literal idiom); CI lint step now BLOCKING (--max-warnings 79 ratchet on
  the remaining warnings); lint added to verify:mechanical

All gates green: 13-package suite, typecheck, LINT, WGSL 51/51, GLSL 6/6,
behavioral 29/29.

+++ packages/pt-webgpu/src/scene/incrementalPatch.ts
@@ -181,7 +181,7 @@ export function canFastPathTopologyResizePatch(
   const curVertexCount = Math.floor(primitive.positions.length / 3);
   const nextPositions =
     'positions' in patch && patch.positions != null
-      ? (patch.positions as Float32Array)
+      ? (patch.positions)
       : primitive.positions;
   const nextVertexCount = Math.floor(nextPositions.length / 3);
